Add rules parameter for selected rules validation
This commit will add extra, optional parameter for csar-validation command.
New parameter is named "rules", and it allows to specify rules that will be validated.
Using this parameter allows to use refrepo/marketplace to validate selected rules and at the same time guarantee that response format will be exactly the same as for validating all rules,
this is in contrast for validating one rule using csar-validate-<rule>, that returns json significantly different then csar-validate.

Signed-off-by: Bartosz Gardziejewski <bartosz.gardziejewski@nokia.com>
Change-Id: I6eb228d360354c0e249032d013e6a23d1f2713ff
Issue-ID: VNFSDK-700